rename cellfinder workflow to avoid circular import issue
