+++ brainglobe_workflows/cellfinder_core/cellfinder.py
