+++ tests/cellfinder_core/test_unit/test_cellfinder.py
@@ -6,16 +6,14 @@ from pathlib import Path
 import pooch
 import pytest
 
-from brainglobe_workflows.cellfinder_core.cellfinder_core import (
+from brainglobe_workflows.cellfinder_core.cellfinder import (
     CellfinderConfig,
     add_signal_and_background_files,
     read_cellfinder_config,
     run_workflow_from_cellfinder_run,
     setup_workflow,
 )
-from brainglobe_workflows.cellfinder_core.cellfinder_core import (
-    setup as setup_full,
-)
+from brainglobe_workflows.cellfinder_core.cellfinder import setup as setup_full
 from brainglobe_workflows.utils import setup_logger
 
 
@@ -29,9 +27,7 @@ def default_input_config_cellfinder() -> Path:
         Path to default input config
 
     """
-    from brainglobe_workflows.utils import (
-        DEFAULT_JSON_CONFIG_PATH_CELLFINDER,
-    )
+    from brainglobe_workflows.utils import DEFAULT_JSON_CONFIG_PATH_CELLFINDER
 
     return DEFAULT_JSON_CONFIG_PATH_CELLFINDER
